Move Leaf/Node-related functionality to nodes.py
